organizing files, destructuring and refactor code
