+++ src/containers/App.js
@@ -1,7 +1,7 @@
 import React, { Fragment, Component } from 'react';
-import CardList from './CardList';
-import SearchBox from './SearchBox';
-import Scroll from './Scroll';
+import CardList from '../components/CardList';
+import SearchBox from '../components/SearchBox';
+import Scroll from '../components/Scroll';
 import './App.css';
 
 class App extends Component {
@@ -27,13 +27,15 @@ class App extends Component {
   }
 
   render() {
-    const filteredRobots = this.state.robots.filter(robot => {
+    const { searchfield, robots } = this.state;
+
+    const filteredRobots = robots.filter(robot => {
       return robot.name
         .toLowerCase()
-        .includes(this.state.searchfield.toLowerCase());
+        .includes(searchfield.toLowerCase());
     });
 
-    if(this.state.robots.length === 0) {
+    if(!robots.length) {
       return <h1 className="tc"> Loading... </h1>
     }
     return (
