+++ src/index.js
@@ -4,7 +4,7 @@ import "./index.css";
 import * as serviceWorker from "./serviceWorker";
 
 import "tachyons";
-import App from "./App";
+import App from "./containers/App";
 
 ReactDOM.render(<App />,
   document.getElementById("root")
